refactoring title to separate widget
